Moved feed models to its own models
