+++ projecto/apiv1/feed/api.py
@@ -5,7 +5,7 @@ from flask.ext.login import current_user
 from kvkit import NotFoundError
 
 from ..hacks import Blueprint
-from ...models import FeedItem
+from .models import FeedItem
 from ...utils import ensure_good_request, project_access_required, jsonify
