+++ projecto/models.py
@@ -18,6 +18,13 @@ import riak
 from settings import DATABASES, RIAK_NODES
 
 
+# Global models.py file.
+#
+# Q: What should go in here?
+# A: Any model that needs to be refered by multiple modules. Or any models that
+#    doesn't have a module associated with.
+
+
 class BaseDocument(Document):
   _backend = riak_backend
 
@@ -88,33 +95,6 @@ class Content(EmDocument):
     elif include_comments == "keys":
       item["children"] = list(Comment.index_keys_only("parent", self.key))
     return item
-
-
-class ArchivedFeedItem(BaseDocument, Content):
-  _riak_options = {"bucket": rc.bucket(DATABASES["archived_feed"])}
-
-  parent = ReferenceProperty(Project, index=True, load_on_demand=True)
-  type = StringProperty()
-
-
-class FeedItem(BaseDocument, Content):
-  _riak_options = {"bucket": rc.bucket(DATABASES["feed"])}
-
-  parent = ReferenceProperty(Project, index=True, load_on_demand=True)
-  type = StringProperty()
-
-  def archive(self):
-    archived_item = ArchivedFeedItem(key=self.key, data=self)
-    archived_item.save()
-    self.delete()
-    return archived_item
-
-  def delete(self, *args, **kwargs):
-    """Overriden as we need to delete the children"""
-    for comment in Comment.index("parent", self.key):
-      comment.delete()
-
-    return Document.delete(self, *args, **kwargs)
 
 
 class Comment(BaseDocument, Content):
